object-recognition-pipeline/requirements.txt
azureml-sdk
Pillow

